[feat] 게시글 수정, 삭제시 본인 확인

+++ src/main/java/com/kakao/cafe/service/PostService.java
@@ -1,6 +1,7 @@
 package com.kakao.cafe.service;
 
 import com.kakao.cafe.domain.Post;
+import com.kakao.cafe.exceptions.UnauthenticatedArticleAccessException;
 import com.kakao.cafe.repository.PostRepository;
 import java.util.List;
 import org.springframework.stereotype.Service;
@@ -26,11 +27,17 @@ public class PostService {
         return postRepository.findByPostId(id);
     }
 
-    public void updatePost(Post post) {
-        postRepository.update(post);
+    public void updatePost(Post updatePost) {
+        Post originalPost = getPostById(updatePost.getId());
+        if (originalPost.getUserId() != updatePost.getUserId())
+            throw new UnauthenticatedArticleAccessException("본인 게시글만 수정할 수 있습니다");
+        postRepository.update(updatePost);
     }
 
-    public void deletePost(int postId) {
+    public void deletePost(int postId, int userId) {
+        Post originalPost = getPostById(postId);
+        if (originalPost.getUserId() != userId)
+            throw new UnauthenticatedArticleAccessException("본인 게시글만 삭제할 수 있습니다");
         postRepository.delete(postId);
     }
 }
